Refactored and fixed specs
