+++ Gemfile
@@ -61,8 +61,7 @@ group :development, :test do
   gem 'brakeman', require: false
   gem 'bundler-audit'
   gem 'byebug', '~> 11.1', '>= 11.1.3'
-  # gem 'debug'
-  # gem 'debase'
+  gem 'debug', '>= 1.0.0'
   gem 'parallel_tests', '3.8.1'
   gem 'readapt'
   gem 'rspec-rails'
